third_party/blink/web_tests/external/wpt/css/css-flexbox/flex-shorthand-calc-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['flex'] = "sign(1em - 1px) sibling-index()" should set flex-basis
  assert_equals: flex-basis should be canonical expected "auto" but got ""
[FAIL] e.style['flex'] = "sign(1em - 1px) sibling-index()" should set flex-grow
  assert_equals: flex-grow should be canonical expected "sign(1em - 1px)" but got ""
[FAIL] e.style['flex'] = "sign(1em - 1px) sibling-index()" should set flex-shrink
  assert_equals: flex-shrink should be canonical expected "sibling-index()" but got ""
[FAIL] e.style['flex'] = "sign(1em - 1px) sibling-index()" should not set unrelated longhands
  assert_true: expected true got false
Harness: the test ran to completion.
